:recycle: Factor out version schema from files app
Moved schema.py into schema/__init__.py and separated Version schema
definitions into schema/version.py so as to break up the single schema
file into individual files based on resource.
